Add client/server prefixes to test classes
